organized files and fixed actions errors
