+++ docs/conf.py
@@ -187,7 +187,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [(master_doc, 'napari-graph-documentation', , [author], 1)]
+man_pages = [(master_doc, 'napari-graph-documentation', [author], 1)]
 
 # -- Options for Texinfo output -------------------------------------------
