+++ src/napari_graph/__init__.py
@@ -1,0 +1,2 @@
+from napari_graph.directed_graph import DirectedGraph
+from napari_graph.undirected_graph import UndirectedGraph
