+++ tox.ini
@@ -1,0 +1,5 @@
+
+[tox]
+envlist = py{38,39,310}-{linux,macos,windows}-{pyqt,pyside}
+isolated_build = true
+toxworkdir=/tmp/.tox
